+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Change Log
 ==========
 
+Version 1.5.1 *(2015-01-??)*
+----------------------------
+
+ * New: `clearHighlightedDates()` - Unhighlights all previously highlighted days.
+ * New: `scrollToDate(Date date)` - Scroll to the month that given date belongs to.
+ * New: `containsDate(List<Calendar> selectedCals, Date date)` - if date is within selectedCals
+ * Fix: When calculating cell width, make sure to cover all of totalWidth.
+
 Version 1.5.0 *(2014-12-15)*
 ----------------------------